Add testAndScorePopulation, use Double in calculating neuron output, not int  and nicer print of the neural net

+++ src/main/java/com/github/stefaniejaeger/neuralnet/network/neuron/HiddenNeuron.java
@@ -28,7 +28,7 @@ public class HiddenNeuron extends Neuron {
     @Override
     public void calculateValue(){
         double e;
-        int xw = 0;
+        double xw = 0;
         //Sigmoid
         for(Connection c : incommingConnections)
             xw+=c.calculateOutput();
